pyproject.toml
[project]
name = "toy-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "google-genai>=1.20.0",
    "python-dotenv>=1.1.0",
]
